one less file

+++ script.sh
@@ -5,8 +5,11 @@ FILE=$1
 
 for dir in $DIRS
 do
+  if [ ! -f "$dir/files/$FILE" ]; then
+    continue
+  fi
   sed -i.bak "s/\/$dir\/files\/$FILE/\/files\/$FILE/g" "$dir/index.html"
   rm "$dir/index.html.bak"
-  [ -f "files/$FILE" ] || ([ -f "$dir/files/$FILE" ] && mv "$dir/files/$FILE" "files/$FILE")
+  [ -f "files/$FILE" ] || cp "$dir/files/$FILE" "files/$FILE"
   [ -f "$dir/files/$FILE" ] && rm "$dir/files/$FILE"
 done
